Fixing compiler warnings.

+++ Core/Datatypes/TriSurfMesh.h
@@ -180,7 +180,7 @@ public:
   {
     Node::array_type nodes;
     get_neighbors(nodes, idx);
-    return nodes.size();
+    return (int)nodes.size();
   }
   int get_valence(Edge::index_type /*idx*/) const { return 0; }
   int get_valence(Face::index_type /*idx*/) const { return 0; }
